Added post-requests and get-requests to yookassa. A handler for incoming requests to the server has been added (by yookassa), the mongodb database is connected, but there is no interaction with it yet.
At the moment, there is a /pay command, where the bot gives a link to the payment (the pay button), after payment, you can check the status of the operation by clicking the check button.

Added processing of messages by the bot after payment, depending on the status of the order.